Fixed config file handling.
When vcproc is run, it tries to ensure the config file uses *NIX line breaks.

Previously, vcproc simply stripped out carriage returns.  If the file was
edited on Mac, this would result in an bad file.  Now it converts carriage
returns to line feeds.  This leaves a file edited in Windows with double
spacing, so a sed line was added to fix that.

Also, vcproc would end up updating the config file's modification time.
This is now no longer the case.