Change `-ss [time] -t [time] -i [file]` to  `-i [file] -ss [time] -to [time]`.
This should fix the difficulty vcproc has had with frame accuracy.  The more
frequent the I-Frames, the more vcproc had trouble starting and stopping where
it was supposed to.  Placing `-ss` and `-t` before `-i` causes ffmpeg to "seek
without playing".  Placing `-ss` and `-to` after `-i` causes ffmpeg to "play
and seek."

The problem was minor with the HD PVR 60 and 0.50-second I-Frames.

The problem became pronounced with the HD 60 S and 0.25-second I-Frames.