Make `trim` save a record of trims.  Enhances commit 5a4d2a1.
The trim process may err.  Example: Running out of storage space.

With no record of the trims the user input, all trims are lost once the
process completes, even when trim encounters errors.

The `trim` process now saves a record of trims in `trim.cfg`.

If `trim.cfg` exists, `trim` offers to load the trims in the file.

The user is free to add more if they so desire.