Fixed `join` cropping all videos.  (Blame commit 48b4c42)
When `join` was modified to use `-filter_complex` instead of
intermediate files, it was intended to apply the video filters to all
joined files.  This way, the source for each file that was to be joined
did not have to prepped in any way.

Due to oversight, this also meant the crop settings were also applied.

Each process that references V_FLTR must now reference CRPFLT if it
intends to work on the file referenced in `vcproc.cfg`.  This enables
join to skip the crop settings for other files.