Automate time code entry for `set` to avoid MPV time code display bug
This has been on the back burner for some time as a convenience
improvement.  However, a recent discovery revealed that MPV does not
always update the time code it displays in the player correctly.
Changing the `osd-level` will cause MPV to update the time code
displayed in the player, which corrects the time code if it is wrong.
This is not an issue in MPV's CLI output.

Now `vcproc` will start `mpv` and redirect its output to
/tmp/vcproc-${USER}/mpv.out

The script will start in "automated mode" and wait for the user to
navigate to the In-Point in MPV.

The user can either click OK to grab the current position from MPV's
output or press ESC to to skip "automated mode" and go into "manual
mode".

The same goes for the Out-Point.

"Manual mode" is the old way of entering In- and Out-Points with the
`vcproc` script.  It now doubles as a way to verify the In- and Out-
Points entered via "automated mode".

The `vcproc` script will now also automatically kill the MPV process it
started.  This is done mainly to avoid having multiple MPV processes
writing to the same output file.  Killing is held off until Audio Peak
Discovery since the user may want to do some additional testing with
the video they are processing.  This is also a convenience improvement.