Fixed sec2tc returning time codes w/ incorrect decimal portions AGAIN
Commit 5914bb8 did not correctly solve the problem.

Incidentally, while it was noted in commit 5914bb8 that the bug did not
affect video processing that used `bc` for calculation, it did affect
the time code display introduced by commit f830bd8 since it used `awk`
instead of `bc`.  So running times displayed in the output of `vcproc`
were sometimes wrong (affected nothing) and are now fixed.